api/requirements.txt
meilisearch
djangorestframework
opencv-python
numpy
pretty-errors
pytesseract